Added dialog windows

+++ src/java/com/bulumutka/polyconstr/models/Edge.java
@@ -4,9 +4,9 @@ public class Edge {
     int id;
     int source;
     int target;
-    int time;
+    double time;
 
-    public Edge(int id, int source, int target, int time) {
+    public Edge(int id, int source, int target, double time) {
         this.id = id;
         this.source = source;
         this.target = target;
